backend/requirements.txt
asgiref==3.2.7
Django==3.0.7
django-archive==0.1.6
django-cors-headers==3.3.0
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
drf-nested-routers==0.91
graphviz==0.14
importlib-metadata==1.6.0
Markdown==3.2.2
pydot==1.4.1
pyparsing==2.4.7
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
zipp==3.1.0
